+++ src/coq_elpi_HOAS.mli
@@ -222,6 +222,13 @@ val isname : RawOpaqueData.t -> bool
 val nameout : RawOpaqueData.t -> Name.t
 val name : Name.t Conversion.t
 
+type global_or_pglobal =
+  | Global of term option
+  | PGlobal of term option * Univ.Instance.t option
+  | NotGlobal
+  | Var
+val is_global_or_pglobal : depth:int -> term -> global_or_pglobal
+
 val in_elpi_modpath : ty:bool -> Names.ModPath.t -> term
 val is_modpath : depth:int -> term -> bool
 val is_modtypath : depth:int -> term -> bool