+++ Changelog.md
@@ -9,6 +9,9 @@
   a Db
 - New `coq.elpi.predicate` to build a term of type `prop` out of a predicate
   name and arguments
+- Change `coq.env.global` now relates a term with a gref, instead of working one
+  way only
+- Change `coq.elpi.accumulate*` generalise clauses over global universe level, and error if algebraic levels are present. It used to warn if levels were present.
 
 ### Vernacular
 - New `Elpi Print` also print the program in `.txt` format